requirements.txt
dash==2.15.0
dash-bootstrap-components==1.5.0
pandas==2.1.3
gunicorn==21.2.0
PyYAML==6.0.1
# Testing dependencies
pytest==7.4.0
selenium==4.15.2
requests==2.31.0